requirements.txt
python-telegram-bot==20.3
fpdf
pillow
aiohttp==3.8.4
